convert pdccommand class to function
